split the tests out
