+++ test/type-error.js
@@ -5,13 +5,16 @@
 var assert = require('assert')
 var the = require('../the-thing-is')
 
-describe('the({ foo:undefined })', function() {
+describe('the({ foo:undefined })', function () {
   describe(".is({ foo:'undef' }", function () {
     it('returns true', function () {
       assert(
         the({ foo: undefined })
         .is({ foo: 'undef' })
       )
+    })
+    it('sets no error', function () {
+      assert.deepEqual(the.last.error, [])
     })
   })
 })
@@ -23,6 +26,10 @@ describe('the({ foo:undefined })', function() {
         the({ foo:undefined }).is({ foo: ['present'] }),
         false
       )
+    })
+    it('sets the right error', function () {
+      console.log('the.last.error: ', the.last.error);
+      assert.deepEqual(the.last.error, [{ foo: ['present'] }])
     })
   })
 })
